refactor: Match 관련 types 분리
