+++ src/dummyMatch.ts
@@ -1,28 +1,3 @@
-export interface MatchCard {
-  matchId: number;
-  city: string;
-  region: string;
-  ground: string;
-  date: string;
-  startTime: {
-    hour: number;
-    minute: number;
-    second:number;
-  },
-  endTime: {
-    hour: number;
-    minute: number;
-    second:number;
-  },
-  cost: number;
-  ageGroup: string;
-  teamId: number;
-  teamLogo: string;
-  teamName: string;
-  teamMannerTemperature: number;
-  sports: string;
-}
-
 export const matchListDummy = {
   data: {
     matchList: [
@@ -78,70 +53,8 @@ export const matchListDummy = {
   },
 };
 
-export interface TeamUser {
-  userId: number;
-  userName: string;
-}
 
-export interface Team {
-  captainId?: number;
-  captainName?: string;
-  teamId: number;
-  teamLogo: string;
-  teamName: string;
-  teamMannerTemperature: number;
-  matchMembers?: TeamUser[];
-  teamUsers?: TeamUser[];
-}
-
-export interface Match {
-  matchId: number;
-  city: string;
-  region: string;
-  ground: string;
-  date: string;
-  startTime: {
-    hour: number;
-    minute: number;
-    second: number;
-    nano?: number;
-  };
-  endTime: {
-    hour: number;
-    minute: number;
-    second: number;
-    nano?: number;
-  };
-  status: string;
-  cost: number;
-  ageGroup: string;
-  sports: string;
-  detail: string;
-  registerTeamInfo: {
-    captainId: number;
-    captainName: string;
-    teamId: number;
-    teamLogo: string;
-    teamName: string;
-    mannerTemperature: number;
-    matchMembers: TeamUser[];
-  };
-  applyTeamInfo?: {
-    captainId: number;
-    captainName: string;
-    teamId: number;
-    teamLogo: string;
-    teamName: string;
-    mannerTemperature: number;
-    matchMembers: TeamUser[];
-  };
-}
-
-export interface MatchDetail {
-  data: Match;
-}
-
-export const matchDummy: MatchDetail = {
+export const matchDummy = {
   data: {
     matchId: 1,
     city: '서울특별시',
@@ -152,11 +65,13 @@ export const matchDummy: MatchDetail = {
       hour: 18,
       minute: 30,
       second: 0,
+      nano:0
     },
     endTime: {
       hour: 20,
       minute: 30,
       second: 0,
+      nano:0
     },
     cost: 30000,
     ageGroup: 'TEENAGER',
@@ -200,7 +115,7 @@ export const matchDummy: MatchDetail = {
   },
 };
 
-export const matchDummy2: MatchDetail = {
+export const matchDummy2 = {
   data: {
     matchId: 2,
     city: '서울특별시',
@@ -211,11 +126,13 @@ export const matchDummy2: MatchDetail = {
       hour: 18,
       minute: 30,
       second: 0,
+      nano:0
     },
     endTime: {
       hour: 20,
       minute: 30,
       second: 0,
+      nano:0
     },
     cost: 30000,
     ageGroup: 'TEENAGER',
@@ -293,19 +210,8 @@ export const matchDummy2: MatchDetail = {
   },
 };
 
-export interface TeamSimple {
-  teamId: number;
-  teamName: string;
-  teamUsers: TeamUser[];
-}
 
-export interface TeamWithUser {
-  data: {
-    teams: TeamSimple[];
-  };
-}
-
-export const userTeamDummy: TeamWithUser = {
+export const userTeamDummy = {
   data: {
     teams: [
       {
@@ -418,22 +324,8 @@ export const userTeamDummy: TeamWithUser = {
   },
 };
 
-export interface WaitingTeam {
-  teamWaitingId: number;
-  teamId: number;
-  teamLogo: string;
-  teamName: string;
-  teamMannerTemperature: number;
-  teamUsers: TeamUser[];
-}
 
-export interface WaitingTeams {
-  data: {
-    waitingTeams: WaitingTeam[];
-  };
-}
-
-export const WaitingTeamsDummy: WaitingTeams = {
+export const WaitingTeamsDummy = {
   data: {
     waitingTeams: [
       {
